Centralize generic type-list helpers
Registration, static registry, and static graph each carried local recursive type-list contains or unique helpers. Moving the generic helpers into type_list.h keeps the policy with the type-list abstraction and deletes the duplicate local metaprogramming.

Constraint: Preserve first-occurrence ordering for unique type lists.

Rejected: Keep registration/static-local helper copies | duplicates generic behavior and makes future changes drift-prone.

Confidence: high

Scope-risk: narrow

Directive: Add generic type-list behavior to type_list.h instead of reintroducing local copies.

Tested: cmake --build build --target dingo_unit_test dingo_matrix_test_construct dingo_matrix_test_construct_invoke dingo_matrix_test_annotated dingo_matrix_test_runtime_container_regressions

Tested: ctest --test-dir build --output-on-failure -R 'dingo_unit_test|dingo_matrix_test_(construct|construct_invoke|annotated|runtime_container_regressions)$|dingo_lit_tests'
